pyproject.toml
[project]
name = "mlzoomcamp-capstone-project"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
  "numpy>=2.1,<3",
  "pandas",
  "matplotlib",
  "jupyter",
  "tensorflow==2.20.*",
  "tensorflow-datasets>=4.9",
  "pillow>=10",
  "scikit-learn>=1.4",
  "tqdm>=4.66",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
  "requests>=2.31",
]
